\name{corpcor-internal}


\alias{positive.svd}
\alias{nsmall.svd}
\alias{psmall.svd}

\alias{pvt.powscor}
\alias{pvt.svar}
\alias{pvt.get.lambda}
\alias{pvt.varlambda}
\alias{pvt.corlambda}
\alias{pvt.check.w}
\alias{pvt.usr}

\alias{print.shrinkage}


\title{Internal corpcor Functions}
\description{
  Internal corpcor functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
